+++ lib/mainImpl.dart
@@ -27,6 +27,7 @@ import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Canvas.dart';
 import 'package:illinois/service/CustomCourses.dart';
 import 'package:illinois/service/CheckList.dart';
+import 'package:illinois/service/Gateway.dart';
 import 'package:illinois/service/MTD.dart';
 import 'package:illinois/service/MobileAccess.dart';
 import 'package:illinois/service/Questionnaire.dart';
@@ -73,6 +74,7 @@ import 'package:rokwire_plugin/service/app_navigation.dart';
 import 'package:rokwire_plugin/service/firebase_core.dart';
 import 'package:rokwire_plugin/service/firebase_crashlytics.dart';
 import 'package:rokwire_plugin/service/local_notifications.dart';
+import 'package:rokwire_plugin/service/places.dart';
 import 'package:rokwire_plugin/service/service.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/app_livecycle.dart';
@@ -154,6 +156,8 @@ void mainImpl({ rokwire.ConfigEnvironment? configEnvironment }) async {
       Assistant(),
       MobileAccess(),
       SkillsSelfEvaluation(),
+      Gateway(),
+      Places(),
     ]);
 
     ServiceError? serviceError = await illinois.Services().init();