+++ lib/ui/RootPanel.dart
@@ -27,6 +27,7 @@ import 'package:illinois/model/Analytics.dart';
 import 'package:illinois/service/Appointments.dart';
 import 'package:illinois/service/Canvas.dart';
 import 'package:illinois/service/Config.dart';
+import 'package:illinois/service/Gateway.dart';
 import 'package:illinois/service/SkillsSelfEvaluation.dart';
 import 'package:illinois/ui/academics/AcademicsHomePanel.dart';
 import 'package:illinois/ui/assistant/AssistantHomePanel.dart';
@@ -35,6 +36,8 @@ import 'package:illinois/ui/athletics/AthleticsTeamPanel.dart';
 import 'package:illinois/ui/canvas/CanvasCalendarEventDetailPanel.dart';
 import 'package:illinois/ui/events2/Event2DetailPanel.dart';
 import 'package:illinois/ui/events2/Event2HomePanel.dart';
+import 'package:illinois/ui/explore/ExploreBuildingDetailPanel.dart';
+import 'package:illinois/ui/explore/ExplorePlaceDetailPanel.dart';
 import 'package:illinois/ui/guide/CampusGuidePanel.dart';
 import 'package:illinois/ui/guide/GuideListPanel.dart';
 import 'package:illinois/ui/explore/ExploreMapPanel.dart';
@@ -58,6 +61,7 @@ import 'package:illinois/service/FirebaseMessaging.dart';
 import 'package:rokwire_plugin/service/events2.dart';
 import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/inbox.dart';
+import 'package:rokwire_plugin/service/places.dart';
 import 'package:rokwire_plugin/service/polls.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/service/service.dart';
@@ -198,6 +202,9 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
       Groups.notifyGroupDetail,
       Appointments.notifyAppointmentDetail,
       Canvas.notifyCanvasEventDetail,
+      SkillsSelfEvaluation.notifyLaunchSkillsSelfEvaluation,
+      Gateway.notifyBuildingDetail,
+      Places.notifyPlacesDetail,
       Guide.notifyGuide,
       Guide.notifyGuideDetail,
       Guide.notifyGuideList,
@@ -211,7 +218,6 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
       HomeFavoritesPanel.notifySelect,
       BrowsePanel.notifySelect,
       ExploreMapPanel.notifySelect,
-      SkillsSelfEvaluation.notifyLaunchSkillsSelfEvaluation,
     ]);
 
     _tabs = _getTabs();
@@ -298,6 +304,15 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
     else if (name == Canvas.notifyCanvasEventDetail) {
       _onCanvasEventDetail(param);
     }
+    else if (name == SkillsSelfEvaluation.notifyLaunchSkillsSelfEvaluation) {
+      _onFirebaseAcademicsNotification(AcademicsContent.skills_self_evaluation);
+    }
+    else if (name == Gateway.notifyBuildingDetail) {
+      _onGatewayBuildingDetail(param);
+    }
+    else if (name == Places.notifyPlacesDetail) {
+      _onPlaceDetail(param);
+    }
     else if (name == Localization.notifyStringsUpdated) {
       if (mounted) {
         setState(() { });
@@ -373,6 +388,9 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
     }
     else if (name == FirebaseMessaging.notifyMapMentalHealthNotification) {
       _onFirebaseMapNotification(ExploreMapType.MentalHealth);
+    }
+    else if (name == FirebaseMessaging.notifyMapStoriedSitesNotification) {
+      _onFirebaseMapNotification(ExploreMapType.StoriedSites);
     }
     else if (name == FirebaseMessaging.notifyMapStateFarmWayfindingNotification) {
       _onFirebaseMapNotification(ExploreMapType.StateFarmWayfinding);
@@ -545,9 +563,6 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
     }
     else if (name == uiuc.TabBar.notifySelectionChanged) {
       _onTabSelectionChanged(param);
-    }
-    else if (name == SkillsSelfEvaluation.notifyLaunchSkillsSelfEvaluation) {
-      _onFirebaseAcademicsNotification(AcademicsContent.skills_self_evaluation);
     }
 
   }
@@ -872,6 +887,24 @@ class _RootPanelState extends State<RootPanel> with TickerProviderStateMixin imp
       if (eventIdValue != null) {
         Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCalendarEventDetailPanel(eventId: eventIdValue)));
       }
+    }
+  }
+
+  Future<void> _onGatewayBuildingDetail(Map<String, dynamic>? content) async {
+    String? buildingNumber = (content != null) ? JsonUtils.stringValue(content['building_number']) : null;
+    if (StringUtils.isNotEmpty(buildingNumber)) {
+      Navigator.push(context, CupertinoPageRoute(builder: (context) =>
+        ExploreBuildingDetailPanel(buildingNumber: buildingNumber)
+      ));
+    }
+  }
+
+  Future<void> _onPlaceDetail(Map<String, dynamic>? content) async {
+    String? placeId = (content != null) ? JsonUtils.stringValue(content['place_id']) : null;
+    if (StringUtils.isNotEmpty(placeId)) {
+      Navigator.push(context, CupertinoPageRoute(builder: (context) =>
+        ExplorePlaceDetailPanel(placeId: placeId)
+      ));
     }
   }